refactor: fix move location filter to correct place
