--- cognite_toolkit/_cdf_tk/client/api/locations/__init__.py
@@ -1,3 +1,0 @@
-from .api import LocationsAPI
-
-__all__ = ["LocationsAPI"]
